refactor(api): organize handlers into separate files
This commit refactors the `internal/api` package by splitting the monolithic `api.go` file into smaller, more focused files. This improves code organization and maintainability.

- `api.go` now only contains the main API handler struct definition.
- `auth.go` contains handlers related to authentication (login, logout, refresh token, JWKS).
- `user.go` contains handlers for user management (create, get, update, delete, change password).
- `helpers.go` contains utility functions used across the API handlers.
